requirements.txt
streamlit
pillow
numpy==1.26.4
pandas
paddlepaddle==3.0.0
paddleocr==2.7.3
